requirements-modal.txt
modal==0.64.0
httpx==0.27.0
feedparser==6.0.11
asyncpraw==7.7.1
pymupdf==1.24.0
pdfplumber==0.11.0
pandas==2.2.0
pydantic==2.9.0
openai==1.50.0
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
pillow==10.4.0
arize-otel
openinference-instrumentation
openinference-instrumentation-openai
opentelemetry-api
opentelemetry-sdk
